pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "callmem"
version = "0.2.0"
description = "Persistent memory for LLM coding agents — local-first, model-agnostic, SQLite-backed"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Dan" },
]
keywords = ["llm", "memory", "coding-agent", "mcp", "ollama", "opencode", "sqlite"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "pydantic>=2.0,<3.0",
    "python-ulid>=2.0,<3.0",
    "click>=8.0,<9.0",
    "httpx>=0.27,<1.0",
    "mcp>=1.0",
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.30,<1.0",
    "jinja2>=3.1,<4.0",
    "cryptography>=44.0",
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.30",
    "ruff>=0.8",
    "mypy>=1.13",
    "pre-commit>=4.0",
]

embeddings = [
    "sentence-transformers>=3.0",
    "sqlite-vec>=0.1",
]

[project.urls]
Homepage = "https://github.com/DANgerous25/callmem"
Repository = "https://github.com/DANgerous25/callmem"
Issues = "https://github.com/DANgerous25/callmem/issues"

[project.scripts]
callmem = "callmem.cli:main"
llm-mem = "callmem.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/callmem", "src/llm_mem"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
